requirements.txt
streamlit>=1.32.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
python-dotenv>=1.0.0
openpyxl>=3.1.2
bcrypt>=4.1.0
plotly>=5.18.0
